refactor: 패키지 구조 구체화(서비스, 컨트롤러, 레포지토리)
