+++ backend/src/main/java/reviewme/keyword/repository/KeywordRepository.java
@@ -1,7 +1,8 @@
-package reviewme.keyword;
+package reviewme.keyword.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
+import reviewme.keyword.domain.Keyword;
 import reviewme.keyword.exception.KeywordNotFoundException;
 
 @Repository
