+++ backend/src/main/java/reviewme/member/controller/ReviewerGroupController.java
@@ -1,4 +1,4 @@
-package reviewme.member;
+package reviewme.member.controller;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
@@ -6,6 +6,7 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RestController;
 import reviewme.member.dto.response.ReviewerGroupResponse;
+import reviewme.member.service.ReviewerGroupService;
 
 @RestController
 @RequiredArgsConstructor
