+++ backend/src/main/java/reviewme/member/domain/ReviewerGroup.java
@@ -1,4 +1,4 @@
-package reviewme.member;
+package reviewme.member.domain;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
