+++ backend/src/main/java/reviewme/review/domain/ReviewKeyword.java
@@ -1,4 +1,4 @@
-package reviewme.review;
+package reviewme.review.domain;
 
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
@@ -10,7 +10,7 @@ import jakarta.persistence.Table;
 import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import reviewme.keyword.Keyword;
+import reviewme.keyword.domain.Keyword;
 
 @Entity
 @Table(name = "review_keyword")
