+++ backend/src/main/java/reviewme/review/repository/ReviewKeywordRepository.java
@@ -1,8 +1,10 @@
-package reviewme.review;
+package reviewme.review.repository;
 
 import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
+import reviewme.review.domain.Review;
+import reviewme.review.domain.ReviewKeyword;
 
 @Repository
 public interface ReviewKeywordRepository extends JpaRepository<ReviewKeyword, Long> {
