+++ backend/src/main/java/reviewme/review/repository/ReviewRepository.java
@@ -1,8 +1,9 @@
-package reviewme.review;
+package reviewme.review.repository;
 
 import jakarta.persistence.EntityNotFoundException;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
+import reviewme.review.domain.Review;
 
 @Repository
 public interface ReviewRepository extends JpaRepository<Review, Long> {
